Add files via upload
This script tests for significant overlap among a set of QTL previously inferred by QTL mapping.
It could be used to test for significant overlap between any set of genomic features that have 
minimum and maximum bounds, but will may be slow if the number as the number of objects increases.
	The script first counts the number of overlap between n QTL distributed across n chromosomes.
This step is done using the GenomicRanges package, starting with a 5 column data frame (see below).
The number of overlaps is then expressed as the average number of overlaps per QTL. If the returned
value is 3, then this means that, on average, each QTL overlaps with 3 others,
	The script then answers the question, does the number of overlaps, differ from what is expected 
by chance? This is done by generating permuted datasets by generating new, random genomic coordinates 
for each QTL, while keeping their observed size. The probability of a qtl falling on each chromosome
is not completely random, but is instead a function of is size (i.e., p(chromi) = length_chromi)/
sum(length_chromi...j)), so that longer chromosomes are more likely to have a QTL assigned to them.
	The average number of overlaps per QTL is calculated for each permutation. The p-value is 
calculated as normal for a permutation test, as (n_permuted > n_observed +1)/(n_perms+1). 9999 perms
was sufficient to estimate a stable p-value for my dataset, and took 6.617476 mins with 27 QTL on a
macbook pro. 
	The script produces a histogram of the permuted values of the overlap statistic, with a red
vertical line that marks the value of the observed value. The upper number on the plot is the p-value
the one below it is the observed average number of QTL overlaps.